reorganize growi-plugin routes
