+++ apps/app/src/server/routes/apiv3/index.js
@@ -1,5 +1,4 @@
-import growiPlugin from '~/features/growi-plugin/server/routes/apiv3';
-import growiPluginAdmin from '~/features/growi-plugin/server/routes/apiv3/admin';
+import growiPlugin from '~/features/growi-plugin/server/routes/apiv3/admin';
 import loggerFactory from '~/utils/logger';
 
 import { generateAddActivityMiddleware } from '../../middlewares/add-activity';
@@ -45,7 +44,7 @@ module.exports = (crowi, app) => {
   routerForAdmin.use('/slack-integration-legacy-settings', require('./slack-integration-legacy-settings')(crowi));
   routerForAdmin.use('/activity', require('./activity')(crowi));
   routerForAdmin.use('/g2g-transfer', g2gTransfer(crowi));
-  routerForAdmin.use('/admin/plugins', growiPluginAdmin(crowi));
+  routerForAdmin.use('/plugins', growiPlugin(crowi));
 
   // auth
   const applicationInstalled = require('../../middlewares/application-installed')(crowi);
@@ -110,8 +109,6 @@ module.exports = (crowi, app) => {
     userActivation.completeRegistrationRules(),
     userActivation.validateCompleteRegistration,
     userActivation.completeRegistrationAction(crowi));
-
-  router.use('/plugins', growiPlugin(crowi));
 
   router.use('/user-ui-settings', require('./user-ui-settings')(crowi));
